requirements.txt
openai>=1.0.0
argparse
python-dotenv>=1.0.0 